+++ src/ui/mainwindow/modelwidget.h
@@ -1,6 +1,6 @@
 /*
  * <one line to give the program's name and a brief idea of what it does.>
- * Copyright (C) 2016 - 2018 Conrad Hübler <Conrad.Huebler@gmx.net>
+ * Copyright (C) 2016 - 2019 Conrad Hübler <Conrad.Huebler@gmx.net>
  * 
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -100,6 +100,7 @@ public:
     void setColorList(const QString& str);
     QString Keys() const;
     void setKeys(const QString& str);
+    QColor ActiveColor() const;
 
 public slots:
     void LoadJson(const QJsonObject& object);