+++ Changelog.md
@@ -1,5 +1,6 @@
 # SupraFit 2.0 pre-Alpha
 
+- data points colors in meta models follow the series color definition
 - biscetion like single parameter optimisation for initial guess of K for NMR and ITC parameters
 - make series names and color changeable in listchart
 - git commit hash is stored in project files